added category related classes.

+++ src/main/java/org/motoc/gamelibrary/repository/implementation/ProductLineRepositoryCustomImpl.java
@@ -3,6 +3,8 @@ package org.motoc.gamelibrary.repository.implementation;
 import org.motoc.gamelibrary.model.Game;
 import org.motoc.gamelibrary.model.ProductLine;
 import org.motoc.gamelibrary.repository.ProductLineRepositoryCustom;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
@@ -14,7 +16,9 @@ import javax.persistence.EntityManager;
 @Repository
 public class ProductLineRepositoryCustomImpl implements ProductLineRepositoryCustom {
 
-    EntityManager entityManager;
+    private static final Logger logger = LoggerFactory.getLogger(ProductLineRepositoryCustomImpl.class);
+
+    private final EntityManager entityManager;
 
     @Autowired
     public ProductLineRepositoryCustomImpl(EntityManager entityManager) {
@@ -24,9 +28,14 @@ public class ProductLineRepositoryCustomImpl implements ProductLineRepositoryCus
     @Override
     public void remove(Long id) {
         ProductLine productLine = entityManager.find(ProductLine.class, id);
-        for (Game game : productLine.getGames()) {
-            productLine.removeGame(game);
-        }
-        entityManager.remove(productLine);
+        if (productLine != null) {
+            for (Game game : productLine.getGames()) {
+                productLine.removeGame(game);
+            }
+            entityManager.remove(productLine);
+            logger.info("Successfully deleted product line of id={}", id);
+        } else
+            logger.info("Tried to delete, but product line of id={} doesn't exist", id);
     }
 }
+
