+++ src/main/java/org/motoc/gamelibrary/repository/implementation/ThemeRepositoryCustomImpl.java
@@ -3,6 +3,8 @@ package org.motoc.gamelibrary.repository.implementation;
 import org.motoc.gamelibrary.model.Game;
 import org.motoc.gamelibrary.model.Theme;
 import org.motoc.gamelibrary.repository.ThemeRepositoryCustom;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
@@ -12,22 +14,27 @@ import javax.persistence.EntityManager;
  * It's the theme custom repository implementation, made to create / use javax persistence objects, criteria, queryDSL (if needed)
  */
 @Repository
-public class ThemeRepositoryImpl implements ThemeRepositoryCustom {
+public class ThemeRepositoryCustomImpl implements ThemeRepositoryCustom {
 
+    private static final Logger logger = LoggerFactory.getLogger(ThemeRepositoryCustomImpl.class);
 
     private final EntityManager entityManager;
 
     @Autowired
-    public ThemeRepositoryImpl(EntityManager entityManager) {
+    public ThemeRepositoryCustomImpl(EntityManager entityManager) {
         this.entityManager = entityManager;
     }
 
     @Override
     public void remove(Long id) {
         Theme theme = entityManager.find(Theme.class, id);
-        for (Game game : theme.getGames()) {
-            game.removeTheme(theme);
-        }
-        entityManager.remove(theme);
+        if (theme != null) {
+            for (Game game : theme.getGames()) {
+                game.removeTheme(theme);
+            }
+            entityManager.remove(theme);
+            logger.info("Successfully deleted theme of id={}", id);
+        } else
+            logger.info("Tried to delete, but theme of id={} doesn't exist", id);
     }
 }
